예매 가능 날짜 조회, 해당 날짜의 좌석 조회, 좌석 예매 요청 및 동시성 UseCase 통합 테스트 작성

+++ chapter2/ticketing/src/main/java/com/example/ticketing/domain/concert/entity/Reservation.java
@@ -34,10 +34,10 @@ public class Reservation {
     private LocalDateTime assignedAt;
 
     @Version
-    private int version;
+    private Long version;
 
     @Builder
-    public Reservation(Long reservationId, Seat seat, String userUUID, int token, AssignmentStatus status, LocalDateTime assignedAt, int version) {
+    public Reservation(Long reservationId, Seat seat, String userUUID, int token, AssignmentStatus status, LocalDateTime assignedAt, Long version) {
         this.reservationId = reservationId;
         this.seat = seat;
         this.userUUID = userUUID;
